Fallos en el Screen Cuenta. Arreglar

+++ app/src/main/java/com/example/tfg/BBDD/Objetos/Cuenta.kt
@@ -3,5 +3,5 @@ package com.example.tfg.BBDD.Objetos
 data class Cuenta (
     var idCuenta: Int,
     var nombreCuenta: String = "",
-    var limiteTotal: Float
+    var limiteTotal: Double
 )
